[Billing] Next Cycle Estimated Balance tests and bug fix (#6435)
This extracts, tests, and documents the code that does the “Next Cycle Balance Estimate”. It also fixes a bug where the promo wouldn’t be displayed if the promo amount was greater than the subtotal.

NOTE: The bug fix means the numbers won't always add up. There's no way to get around this, without changing the design.

Also, I thought of a scenario we weren't handling for. If a new customer has a promotion, but hasn't accrued charges yet, their uninvoiced balance would be 0, which means we wouldn't display the promotion at all according to the previous logic. I changed this, so that if the subtotal >= 0 we display the promotion.

* Extract and test billing fns

* Extract, test, and clean

* Fix comment

* Cleanup

* Handle for balanceUninvoiced == 0

* Always show promo if there is one