Update linode/disk.reducer and actions to use MappedEntityState(2) (#6408)
* Update linode/disk.reducer and actions to use MappedEntityState(2)

* Add link to Longview form disk storage display

* Change "Storage" to "Allocated"

For Linode disks, we show a Storage panel (LinodeDiskSpace) that shows
the total amount of allocated vs. free storage, along with a note
to use df -h to see actual disk usage. This has proven to be
confusing to customers, who see "Storage", see that everything is
allocated (as it will be for all new Linodes deployed from an image),
and panic because they think their disks are full.

Per recommendations, change the wording to show "Allocated" instead
for clarity.

* Refactor getAllLinodeDisks request to use createRequestThunk