Dismiss TPA modal when external link clicked (#6455)
Small adjustment here. This dismisses the TPA modal when "Go to GitHub" is clicked. The ideal flow is a complete circle in the same tab, but until then, this at least makes it so that the modal isn't still open when you navigate back to your Cloud Manager tab after enabling TPA.