List regions that support GPU in the select plan helper text
- Add a utility to turn a list of strings into an Oxford comma string
("Denver, Boulder, and Roanoke")
- Add regions to SelectPlanPanel and if the GPU tab is disabled
list the regions where it won't be disabled.

I'm not happy with how I did this, but I feel that it's worth it:
if a user selects a region, tries to select a GPU plan, and just sees
"This plan is not available in this region" they're left to use trial
and error to figure out what works.