Revert setting lastUpdated on successful create
I wasn't sure about this choice, and setting lastUpdated in the
addMany helper method broke an existing test for the nodeBalancer.reducer.
On balance let's not do this.

I also used the onGetPageSuccess helper in the NB reducer (it wasn't
available when that reducer was refactored.) The behavior should be
unchanged.