Review feedback
- Update nodeBalancer.selectors to not crash with the new pattern
- Remove unneeded updateNodeBalancerConfigs request. This was needed
to clean up the configs in the store when they were all a single blob.
Now that they're separated by NB this isn't as big an issue, and we
can just request the configs for that NB and end up with an accurate
state.
- Removed { results: 0 } from ensureInitializedNestedDefaultState

I'm going to request that the API team add a secondary event for
config creation and deletion. It would simplify the logic, and for
deletes at least it would enable us to avoid making a request from
nodeBalancers.events.ts.