Allow clearable numeric inputs in DomainRecordsDrawer
We were using +e.target.value to coerce string inputs to numbers.
This allowed our types to play well together internally, but it had
the drawback of not allowing users to clear the input, since '' coerces
to 0.

The solution is pretty messy: we stop the coercion in the event handlers,
and instead handle conversion just before sending data to the API.

I had to update the typings in the drawer, and its parent, to make
this work. It is very messy, and this whole workflow is screaming
for a refactor.