Fix editing Domain records
This was broken in develop. I think it's due to the SDK refactors;
the schema for updating records was incorrect and had been for a while,
since type isn't an accepted or required field when editing. I'm
guessing the refactors applied the schema in a way or time when it wasn't
being applied previously.

I removed type() from updateRecordSchema, which resolves the issue.