Fix @webpack/svgr settings
Make our prefixing config more explicit, so that class names are not
auto-prefixed but ids are. We do some CSS targeting on class names
in SVGs, which prefixing breaks. Otoh id conflicts were causing
a visual bug in the Linode create flow, where the Mumbai flag svg
conflicted with the circle svg in the select plan panel.