Add feature flags loaded state to Redux
There was a flaw with my last approach, which was that there was a white screen gap between the splash screen and the content of the main app.

By adding the feature flag loading state to the Redux store, SplashScreen can now listen for this state and render the splash screen until the flags are loaded.