Handle host maintenance
The API has added a new status for Linodes, "stopped", which will
be the case when a Linode's host is undergoing maintenance. The API
will disallow certain actions in this case, but we should a) notify
the user and b) prevent them from submitting these actions in the
first place.

The actions are:
- rebooting
- booting
- rescue
- resize
- rebuild
- shutdown
- clone