Wait to render main content until feature flags are loaded (#6488)
This makes it so that the splash screen is shown until feature flags are loaded from LD. This is necessary for the large changes we're going to be making (like moving the PrimaryNav), otherwise the app would jump around once the flags are loaded.

To do this, I haded a field to our Redux store so `<SplashScreen />` could listen for changes made in `<IdentifyUser />`. I used a similar pattern to the existing "initialLoad" slice of Redux state.

* M3-4195 Move PrimaryNav (#6483)

This moves the PrimaryNav from the left side of the viewport to the top.

* CMR files

* Wrap App with featureFlagConsumer

* Move PrimaryNav to top

* Remove unnecessary props

* Remove useLocation

* Hide main content until flags loaded

This was much easier than I thought. <App /> already had `flagsLoaded` state, set by <IndentifyUser />. This makes it so that MainContent isn't rendered unless the flags are loaded.

* Add feature flags loaded state to Redux

There was a flaw with my last approach, which was that there was a white screen gap between the splash screen and the content of the main app.

By adding the feature flag loading state to the Redux store, SplashScreen can now listen for this state and render the splash screen until the flags are loaded.

* Fix test

* Undo json formatting