M3-4239 CMR table base (#6538)
This contains the "base" for CMR work on tables. I'm not exactly sure all of these components will be needed, but I included all of our components that deal with tables.

These new files are copy/pastes of the original. As a reminder, the idea is that CMR work is done in these new files/components, which are conditionally rendered:

```tsx
flags.cmr ? <Table_CMR /> : <Table />
```

When CMR is complete/deployed, we'll copy paste in reverse... (copy the CMR files and paste the contents in the original, to maintain some Git history).

* CMR table base

* Fix linting errors