M3-4087 Update setDeletion logic for Events
To determine if an event was related to an entity that has been
deleted, we were checking (among other things) that the relevant
deletion event had a status of finished or notification. This caused
a slight problem when deleting a Linode, since a linode_shutdown
event is generated at exactly the same time as the linode_delete
event, and it was possible to click on this event in the events menu
before the delete had finished; however, the API already considered
this deleted (or maybe it was just our Redux store?), so clicking
the event caused a 404.

I adjusted this logic so that we only check that the delete event
didn't fail (status !== failed). This handles the case above,
and should prevent the behavior the original logic was (I think)
intended to avoid: a long-ago failed deletion event should not
render all subsequent events pointing to that entity un-clickable.