Fix sort order on Domains Landing (#6552)
Originally reported in https://github.com/linode/manager/issues/6550. It looks like this was a regression from the EntityTable component.

To fix, I added an optional `initialOrder` prop to `<EntityTable />`, which is passed down to the `order` and `orderBy` props on `<OrderBy />`. 

This new prop is optional since `<OrderBy />` defaults them to `label` and `asc` anyway.

I also added a test to ensure default order (as well as a helper function for testing order of elements).

* Allow initial order on EntityTable

* Add unit test

* Review feedback