Move mock toggling to Webpack
Instead of importing the worker directly into index.tsx, call it
from where it's defined in testBrowser.ts. Then add this file as
an entrypoint if and only if we're in development mode. The actual
mocks are determined by the same .env variable as before; if the
variable is true, the full set of mocks is loaded. If not, an
empty array is passed.