Add: API Mocks with mock-service-worker (#6590)
* Remove firewallDevice api mocking and install mws library

* Working test mocking

* Add browser mocking

* Clean up factories and factory exports

* Add more mock endpoints

* Mock user preferences call

* Mock more Linode things

* Mock stats and transfer

* Comment out service worker initt

* Mock /kubeconfig

* Update docs

* Linter

* Refactor VolumeTableRow

* Clean up

* Fix authentication settings test

* Remove stray 'only' from firewall row test

* Fix profile mocking

* Reorg into src/mocks

* Move mock toggling to Webpack

Instead of importing the worker directly into index.tsx, call it
from where it's defined in testBrowser.ts. Then add this file as
an entrypoint if and only if we're in development mode. The actual
mocks are determined by the same .env variable as before; if the
variable is true, the full set of mocks is loaded. If not, an
empty array is passed.

* Do it in a less stupid way

* Re-add testServer.js, which unfortunately had the same name as a file I added

* Review feedback

- Remove duplicate testServer.js file
- Add worker.stop() to deactivate the service worker if
mocking is not enabled (this doesn't seem to delete or unregister the worker,
which happens when all relevant tabs are closed).
- Update logic in src/constants to explicitly check for 'true'
- Update date format so that Luxon doesn't break

* Refactor worker init logic