Move large account logic
Added an optional callback to getAll; used to invoke userPreferences
logic and set and account as is_large_account. Has to be done this way
because:

1. getAll is imported in the store (often), so dispatching an action
from within it would be a circular dependency
2. Can't use logic in the .then() when calling getAll, as I did initially,
since a huge account will crash before reaching the then block. This way,
as soon as the first page of a getAll request is complete, the callback will
be fired off. The worst that will happen in this case is that a user
has to reload the page (because again it crashed), at which point
the new preferences will be requested and the large account handling
should kick in.