M3-4332 Change: Add is_large_account to user preferences (#6663)
* Add hook, user preference, etc. for classifying large accounts

* Update search bar logic

Update search logic so that:
- useReduxLoad doesn't request account data for large accounts
- The search bar uses either API search or client search,
depending on the size of the client, but never tries to do both

* Update search landing page for large accounts

* Tweak preferences logic and mocks

* Use page_size param when searching via API

* Move large account logic

Added an optional callback to getAll; used to invoke userPreferences
logic and set and account as is_large_account. Has to be done this way
because:

1. getAll is imported in the store (often), so dispatching an action
from within it would be a circular dependency
2. Can't use logic in the .then() when calling getAll, as I did initially,
since a huge account will crash before reaching the then block. This way,
as soon as the first page of a getAll request is complete, the callback will
be fired off. The worst that will happen in this case is that a user
has to reload the page (because again it crashed), at which point
the new preferences will be requested and the large account handling
should kick in.

* Review feedback/fix tests

- Remove extraneous props.search call in SearchLanding's useEffect
(was causing a bunch of unnecessary searches and breaking a unit
test)

- Add comment explaining why the cb argument to getAll was needed

* Feedback part 2

- If a is_large_account: true app goes below 500 entities (in this case
Domains), call updatePreferences to set the large account var to false