Utility/Easter Egg: user preferences editor (#6668)
Since user preferences are tied to OAuth Clients, to change Cloud Manager preferences you must grab the short-lived token from the app and use that to curl.

This adds a secret "user preference" editor directly in Cloud, accessible by going to /profile/settings?preferenceEditor=true. The editor is pretty simple; it allows you to enter arbitrary JSON and submit it (validating that it's valid JSON first). This makes it easier to quickly edit preferences when developing features that depend on it.

* Add preference editor

* Use Link component so doc opens in new tab

* Fix dialog styling

* Add documentation

Co-authored-by: Jared Kobos <jaredkobos@gmail.com>