Custom dev tools (#6655)
Following the lead of https://kentcdodds.com/blog/make-your-own-dev-tools.

This adds "custom dev tools" to Cloud. So far it only supports feature flags (and only CMR). To do this, I created a new slice of state, mockFeatureFlags, which the dev tools dispatches actions to when flag toggles are checked.

Both feature flag providers – the HOC and the hook – subscribe to this state and override the real feature flags with the mock feature flags.

The method of enabling the dev tools is the same from the blog post. They are enabled by default when running locally, and disabled by default when the prod bundle is built.

This means the dev tools will be accessible in production by adding ?dev-tools=true to the URL. I think this is fine useful, but we should be mindful of what's allowed in them. There's a concept of importing "local", untracked dev tools present in the code which will allow us to add tools to switch environments, for example, which is something we'd want locally but not expose in production.

* Custom dev tools

* Fix tests

* Add gitignore and documentation