M3-4373 Confirm enabling backups (#6738)
* M3-4373 Confirm enabling backups

When enabling backups from either the Linode action menu or Linode
detail /backups tab, we didn't ask for confirmation. Since this is a billing
change, we should show a confirmation modal in these cases. This is especially
true now that we're using Reach components for the action menu.

This was supposed to be a small change, but our pattern here was janky. When a user
clicked the action menu for Enable Backups, we were:

1. Redirecting to /backups
2. Passing router state such that
3. in the cDM of LinodeBackups, we called enableBackups() automatically
if the state had been passed.

In the new pattern, we'll use a single confirmation dialog, reused in both locations.
When the action menu is clicked, it will:

1. Open the dialog
2. The dialog can be submitted, enabling backups
3. Once that's complete, redirect to /backups as before.

This is part 1, which adds the confirmation modal to the /backups tab.
It breaks the action menu, which will have no effect (other than redirecting
to the /backups tab) until part 2 is merged.

- Broke out the Placeholder and EnableConfirmationDialog into separate components
- Allowed the dialog component to manage its own state and API calls, so that
it can also be used from the action menu.

* Fix typo