M3-4310 Flavor text in LinodeNews banner
Designs for the link to the most recent Cloud changelog on the dashboard
included possible "flavor text" summarizing the release changes. Decided
that the best way to do this is through LaunchDarkly. If the version
of the changelog flag matches the currently deployed version, the notes
are shown. If the flag is off, or the versions don't match, nothing
is displayed. This way, the display is safe: if we forget to write
a message for a given release (which seems very likely), nothing will
go wrong and no incorrect information will be displayed.

The downside of this is that we're once again relying on LD as a
pseudo CMS. Especially given the cost of the service, we shouldn't
be increasing our reliance on it for non-flag-related things.