Always persist group by tag
I introduced a change in #6594 which meant that the “Group by Tag” toggle setting would be ignored if there are Linodes with pending maintenance. This was to ensure Linodes with maintenance always appeared at the top of the list.

We’ve gotten a lot of feedback from customers that are unhappy with this behavior, and in retrospect, I can see how this is frustrating.

This reverses that behavior. I figure with the upcoming changes to the CMR dashboard, Linodes with pending maintenance will be plenty visible.