Fix handling of empty IPv6 responses
If a user uses the API to manually set their Addresses field for a rule
and leaves either ipv6 or ipv6 undefined (instead of an empty array, which
is how we handle this), the app will crash when trying to open the edit drawer.

Not sure that `undefined` should be a valid value here, but since the API allows
it for now we should handle the edge case.