Fix tab change handler (#6824)
This fixes an issue with the StackScript Landing tabs. Two things were happening:

The handleTabChange function was happening on every keystroke from the search bar, resulting in a “Cannot read property 'category' of undefined” error (reported via Sentry).
The query string was not being properly updated when switching tabs.
These things were happening because the handleTabChange prop was incorrectly being passed to <TabPanels /> instead of <Tabs /> in TabbedPanel.tsx.

This also fixes the fact that the Select Plan panel wasn’t resetting in the LKE “Add a pool” drawer when switching tabs.