M3-4383 Always honor Linodes "group by tag"(#6821)
I introduced a change [here](https://github.com/linode/manager/pull/6594/files#diff-117f855269a2f760c3c15da2fb97afd4R307) which meant that the “Group by Tag” toggle setting would be ignored if there are Linodes with pending maintenance. This was to ensure Linodes with maintenance always appeared at the top of the list.

We’ve gotten a lot of feedback from customers that are unhappy with this behavior. In retrospect, I can see how this is behavior is frustrating.

This reverses that "feature". I figure with the upcoming changes to the CMR dashboard, Linodes with pending maintenance will still be plenty visible.