M3-4477 fix: Managed Issue Drawer crash(#6850)
The Issue History drawer was crashing on prod test account 4. This was happening because the timezone on this account is set to "" (empty string). The useTimezone function in use by the drawer gets the raw value (defaulting to GMT if undefined), and didn't make sure it was a valid timezone before parsing, hence the crash.

To fix this I modified useTimezone to call getUserTimezone(), which ensures the timezone is valid. Feel free to disagree with this behavior. A case could be made for useTimezone returning whatever the user has set, though it seems more useful to ensure the zone is valid. To get the raw value, a consumer could always use useAccountManagement to get profile data directly.