Fix "Cannot read property 'searchText' of undefined" error (#6823)
This error was being reported via Sentry. AFAICT this was happening when a user clicks the error message in the Search Bar. It doesn't crash the app or anything, but this will clean up some noise.

Zooming out, it looks like some of the search errors were happening when a restricted users on a large account (and thus, using the API search) uses the search bar. In this case, LKE requests return a 403, resulting in an API error, making the search feature unusable. This also includes a (hacky) fix for that.