M3-4452 Upgrade sanitize-html (#6839)
Previously there was an issue where some HTML tags were removed instead of escaped, which led to confusing support ticket bodies (see ticket for more details). This was happening even though we have sanitize-html set to "escape".

It looks like this was a bug in that library. Upgrading the library fixed the issue, with no regressions as far as I can tell.