Deprecate Atlanta region
Since it's unclear how much of this deprecation process will
be consistent in the future, this is a mostly manual attempt
to prevent users from accessing the Atlanta datacenter.

It manually checks the region id inside the RegionSelect, and if
it matches Atlanta it marks the option as disabled.

- Add a tooltip with explanation text and a link to the blog post
- Add a check that only disables the option if the user doesn't
already have Linodes in Atlanta