Deprecate Atlanta region
Since it's unclear how much of this deprecation process will
be consistent in the future, this is a mostly manual attempt
to prevent users from accessing the Atlanta datacenter.

It manually checks the region id inside the RegionSelect, and if
it matches Atlanta it marks the option as disabled.

- Add a tooltip with explanation text and a link to the blog post
- Add a check that only disables the option if the user doesn't
already have Linodes in Atlanta

Add fake Atlanta to response

The API has updated their approach to Atlanta deprecation.
Any user who is able to deploy resources to Atlanta will have
a customer tag; users with that tag will see us-southeast in the
response from /regions. Otherwise, us-southeast will be omitted.

In Cloud, this means that if us-southeast is returned in the regions
array, the customer can deploy things to ATL normally and there's nothing
special for us to do. If ATL is not in the response, we want to show
messaging to avoid confusing people. In this case, we add a fake
Atlanta to the regions response, and make that option in the RegionSelect
disabled. It can't be selected or submitted, but includes hover text with
a link to the relevant blog post.