Remove event handling in CMR action menu
Our old ActionMenu, which was based on MUI, triggered underlying
events when an action in the menu was clicked. This caused some
issues, which we fixed by adding e.preventDefault() and (sometimes!)
e.stopPropagation() in our action handlers in individual actionmenu
consumers.

This logic was carried over into the new ActionMenu, which uses
Reach UI and doesn't seem to fire these events. According to their
docs (at https://reach.tech/menu-button/#menulink-onselect):

This won't work because we actually do not call the onClick handler to activate Menu or select MenuItem or MenuLink components. This is because a user may mouse down on a menu button, drag over a menu item and then select it by releasing the mouse trigger. A user may also start clicking one item, then drag to another item before mouseup to change their selectiion.

One consequence of this difference is that when an action is triggered,
our existing preventDefault() logic prevents Reach from doing its thing,
which in this case includes closing the action menu. MUI took an onClose()
as a prop, which we were calling from many individual actions to resolve this.

Removing our default prevention logic causes the menus to function correctly.