Use MenuItem instead of MenuLink
The MUI-to-Reach refactor used the MenuLink component for our actions,
even though none of our actions are true links (in the sense of providing
an href). This was done because the interface most closely matched the MUI
pattern we were used to. This turns out to be a bit of a hack that could
have more consequences besides the menu close issues in the previous commit.

The correct thing to do (I think) is use MenuItem and provide an onSelect
handler instead of onClick. onSelect will close the menu and execute our
function, without calling any of the actual onClick handlers of e.g. a button
or <a>.