M3-4517 followup: Change URL on Linode modal close (#6867)
Currently on devlop, when you navigate to e.g. /linodes/:id/resize, you are taken to the Linode Details page with the resize modal opened. When you close the modal, the URL is unchanged, so if you refresh the page the modal is opened again.

This makes it so the URL is updated to remove the /resize bit.

I also renamed a variable/fn for clarity.