M3-4506 Close action menu after action is selected (#6871)
* Remove event handling in CMR action menu

Our old ActionMenu, which was based on MUI, triggered underlying
events when an action in the menu was clicked. This caused some
issues, which we fixed by adding e.preventDefault() and (sometimes!)
e.stopPropagation() in our action handlers in individual actionmenu
consumers.

This logic was carried over into the new ActionMenu, which uses
Reach UI and doesn't seem to fire these events. According to their
docs (at https://reach.tech/menu-button/#menulink-onselect):

This won't work because we actually do not call the onClick handler to activate Menu or select MenuItem or MenuLink components. This is because a user may mouse down on a menu button, drag over a menu item and then select it by releasing the mouse trigger. A user may also start clicking one item, then drag to another item before mouseup to change their selectiion.

One consequence of this difference is that when an action is triggered,
our existing preventDefault() logic prevents Reach from doing its thing,
which in this case includes closing the action menu. MUI took an onClose()
as a prop, which we were calling from many individual actions to resolve this.

Removing our default prevention logic causes the menus to function correctly.

* Use MenuItem instead of MenuLink

The MUI-to-Reach refactor used the MenuLink component for our actions,
even though none of our actions are true links (in the sense of providing
an href). This was done because the interface most closely matched the MUI
pattern we were used to. This turns out to be a bit of a hack that could
have more consequences besides the menu close issues in the previous commit.

The correct thing to do (I think) is use MenuItem and provide an onSelect
handler instead of onClick. onSelect will close the menu and execute our
function, without calling any of the actual onClick handlers of e.g. a button
or <a>.

* Add React.memo to the Linode table

* Can't use Jest in Storybook