Fix console error (isResponsive prop not recognized)
Our non-CMR Table component had an optional isResponsive prop.
The CMR version does not have this prop, though it was copied over
in the props interface and included in many instances of the table.

However, since it isn't used anywhere, it ends up being passed down
to the underlying component as an html attribute, which triggers a
React console warning.

Removed isResponsive from the CMR table interface and used that
to hunt down all the CMR tables that were using it.