Extract data from Axios response globally
In the JS client, virtually every method includes a `.then(response =>
response.data)` to extract the actual API response from the Axios wrapper.
This led to problems in the few cases where we forgot to do this.

Moving the extraction to request.ts ensures that we never forget; I
changed the type of the request object to return a Promise<T> instead
of AxiosResponse<Promise<T>>. This means that TypeScript should pick
up on any place in Manager that tries to use the incorrect access pattern,
and won't allow client methods to do the response.data thing. Also saves
a lot of typing and boilerplate.