Filter firewall rules inside OrderBy
Previously, modified and unmodified firewall rules were filtered
into separate lists in the FirewallRulesTable, and only unmodified rules
were passed to OrderBy. This was so that modified rules appeared unsorted
at the bottom of the list (in the order in which they were added).

This worked until recently, when changes to OrderBy caused a very
confusing react key collision. To avoid this, this commit moves
the filtering to after the OrderBy sorting.