M3-4521 Handling for unknown types (#6888)
It’s possible for a Linode to have a type that is not included in the responses from /linode/types or /linode/types-legacy. This is usually do to specific upgrades or other special circumstances. The only way to get info for these types is to request it from /linode/types/:id.

Currently, we don’t handle this very well; we display stuff like “Unknown Plan” and fail to show correct pricing information in the Backups Drawer when enabling backups.

This does two things:

1. Make individual requests for “shadow plan” types if there are Linodes with unknown plans (via Redux).
2. Adjust the way components access types data from Redux, given point the previous point.