M3-4529 fix OrderBy lifecycle bug (#6903)
Fixes a bug in OrderBy that led to the FW regression yesterday. This makes use of refs and does the equality check in the function body itself instead of useEffects, which are run after render.

While I was here I again tried to get rid of this equality check with no luck. I tried storing previous values of order and orderBy but this becomes very complicated (you need to keep track of previous values on a stack) and I'm not sure the tradeoff is worth it.

We do have performance issues when sorting large lists, but AFAICT the issues come from the massive DOM tree and not the actual equals() calls.

Co-authored-by: Jared Kobos <jaredkobos@gmail.com>
Co-authored-by: Kayla Wilkins <kwilkins@linode.com>