M3-4534 Inject EUUID to /profile from HTTP header (#6918)
Currently we read a user's EUUID from `/account`. Unfortunately this endpoint is not available to restricted users, so the API will start returning an `X-Customer-Uuid` HTTP header with this value.

This adds a `_euuidFromHttpHeader` field to our `ExtendedProfile` type, (which is injected by Axios middleware) so that this value can be used for non-restricted and restricted users alike.