Finalize logic
When creating a VLAN:

1. If the user selects a region first (normal pattern), filter
the Linode select by that region.

2. If the user leaves region blank and first selects a region,
automatically select that Linode's region (and consequently filter
the Linode select by that region, so subsequent Linodes are all
in the same region).

3. If a Linode is selected and the region is toggled, clear the
selected Linode if it's not in the newly selected region. (Since
we auto-select the region when the Linode is first selected, any
toggle of the region input can be assumed to be in a different
region than the currently selected Linodes).

4. Reset the checkbox (for reboot after create) whenever the list
of Linodes is reset or manually cleared for any reason.