M3-4543 Add: VLAN create modal (#6934)
* Fix key warning

* Add form fields to create modal

* Add VLAN context for opening/closing modal; move modal to main content level

* Clean up

* Add initial unit tests

* Add updated cache for regions data

* Error handling and tests

- Fix/finish unit tests
- Add Notice for displaying general errors
- Clear form when the modal opens

* Review feedback #1

- Adjust spacing in create modal
- Don't set a default region (we don't do that elsewhere and
our alphabetical-by-country sorting makes it hard to dynamically
choose a stable default)
- Adjust schema now that description has a limit of 64 chars

* Clear checkbox when Linode select is cleared

* Make dialog context generic

* Fix Linode filtering by region

- Update LinodeMultiSelect to allow it to optionally
be controlled from its parent. (it now takes selectedLinodes,
a number[] which matches the signature of handlechange).

- In Vlan create, if no region has been selected, filter Linodes
by all allowed regions (any region with capabilities.vlans). If a
region is selected, a) clear the selected Linodes and b) filter
the Linode select by the target region.

* Select a Linode's region when selecting a Linode (and region hasn't already been selected)

* Finalize logic

When creating a VLAN:

1. If the user selects a region first (normal pattern), filter
the Linode select by that region.

2. If the user leaves region blank and first selects a region,
automatically select that Linode's region (and consequently filter
the Linode select by that region, so subsequent Linodes are all
in the same region).

3. If a Linode is selected and the region is toggled, clear the
selected Linode if it's not in the newly selected region. (Since
we auto-select the region when the Linode is first selected, any
toggle of the region input can be assumed to be in a different
region than the currently selected Linodes).

4. Reset the checkbox (for reboot after create) whenever the list
of Linodes is reset or manually cleared for any reason.

* Fix tests