M3-4563 Fix: Search results not displaying for some restricted users
When a user has >500 Domains, we trigger our useApiSearch() logic,
to prevent the search bar from crashing. A corner case was recently
discovered that broke this logic:

If a restricted user has access to >500 Domains, the search bar
returns an empty results list inline, and the search landing page
displays an error.

This is because restricted users, regardless of access/grants, always
receive a 403 when attempting to access /lke/clusters. This may change
at some point, but is not expected until at least 2021.

Since we were making API search requests to /lke/clusters unconditionally,
this meant that we always ended up in the .catch() block of our search
logic, and results were not displayed. (Promise.all() bails as soon as
a single request fails).

By making this particular request only for nonrestricted users, we can
avoid this situation. I return Promise.resolve([]) for restricted users
for Kubernetes, which loses some UI (before we showed an error message
on the landing page saying "couldn't retrieve results for: Kubernetes",
which will now not work at all for large accounts).

I also added a conditional to useReduxLoad in the search landing page,
which was unconditionally loading all entities on page load even
for large accounts.