Change: Request Domains on load to determine account size
Since the API has determined that a "thing count" endpoint
is not feasible, and our existing logic (which waits for
a getAllDomains request to be triggered, then saves is_large_account
to user preferences if > 500 results come back) has caused
frustration in customers from huge accounts who don't like the
initial crash of their accounts, we need to update our approach.

I added a new Redux state, accountManagement, to the __resources
envelope. This is debatable; we could have stored this data in
a context, or added it to an existing state component. I chose
this way because it's fairly clear and our Redux structure makes
it easy to bring in thunks from other parts of the state.

In practice, the checkAccountSize thunk is called on app load.
It requests a single page of Domains (page size of 100 to keep
the request fast, although it probably doesn't matter that much).
If the response.results from that request tells us that there
are a higher number of Domains than our threshold (currently 500),
it dispatches an action that marks the account as large in Redux.
We then read this value, rather than one from user preferences,
to determine whether to use API or redux-based pagination/search.

In addition, once we've made this request, the getDomainsPage
thunk we're using will check if the total number of domains
has been returned by the single request (this is existing
behavior for this thunk). In this way, for an account with few
Domains, the request we make on initial load obviates the need
for further requests when navigating to /domains or searching.

Finally, I removed all the existing user preferences logic. Since
we need to make this request when the app loads anyway (or make it
conditional based on whether we already have a value in preferences),
going through a separate endpoint is an unnecessary detour. This also
makes it simple to handle the case of an account that has 500+ Domains
and then deletes ~100, bringing them under the threshold.